+++ anyblok_wms_base/bloks/wms_core/goods/__init__.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-# flake8: noqa
 # This file is a part of the AnyBlok / WMS Base project
 #
 #    Copyright (C) 2018 Georges Racinet <gracinet@anybox.fr>
@@ -8,18 +7,15 @@
 # v. 2.0. If a copy of the MPL was not distributed with this file,You can
 # obtain one at http://mozilla.org/MPL/2.0/.
 
-from . import split
-from . import aggregate
-from . import splitter
-from . import arrival
-from . import move
-from . import unpack
+from . import goods
+from . import type as gtype
+
+# even with 'import as', there is masking of the builtin
+# (at least on Python 3.5). Let's remove it if it's there
+# to avoid leaving a trap here
+globals().pop('type')
 
 
 def reload_declarations(reload):
-    reload(split)
-    reload(aggregate)
-    reload(splitter)
-    reload(arrival)
-    reload(move)
-    reload(unpack)
+    reload(goods)
+    reload(gtype)
